requirements.txt
praw
openai
pymongo
dnspython
flask
python-dotenv
requests
